introduced cron scheduling, implemented singleton Scanner class and moved all html processing and job object creation there
