+++ src/main/resources/application-dev-sample.properties
@@ -2,3 +2,5 @@ spring.datasource.url=jdbc:postgresql://localhost:5432/indeed-db
 spring.datasource.username=${POSTGRES_USERNAME}
 spring.datasource.password=${POSTGRES_PASSWORD}
 server.port=5001
+
+cron.schedule=0 5 * * * *
