+++ src/main/resources/application.properties
@@ -1,6 +1,5 @@
 server.port=${PORT:5000}
-update.delay=PT4H
-initial.update.delay=PT10S
+cron.schedule=${CRON}
 spring.task.scheduling.pool.size=10
 
 spring.profiles.active=${SPRING_PROFILE}
